Restructured how tests are ran. Began using libcurl
I restructured how tests are ran/written. There is a singular
main file that runs all tests recognized by the testing framework.
Tests are now written as header files and are placed in
tests/include/

I've also added the beginnings of using libcurl. I need to add
documentation for the library as a dependency and also need to
write a test to make sure it works during build.