Added API key as argument for IngredientRequest class
I've added the API key as an argument for the IngredientRequest
class. This stems from wanting to keep the API key separate from
the class implementations. I may go back and centralize the API
key more to increase security but this isn't a huge concern for
me since I'm not going to commercialize this software.