Finalized Nutrient map in IngredientResult class
Finalized #20 by cleaning up the population of the map in the
IngredientResult constructor. The issue was with passing the
rapidjson::Value reference as a non-const and then creating a const
reference of this in the function. I'm not quite sure why that caused
the issue, but that was certainly it. I haven't been able to able to
create a reference in the Nutrient constructor now that I'm passing
the rapidjson::Value reference as a const but I'm not too worried
about it since it hasn't impacted speed.