Add IngredientResult constructor with file handle
I've added a constructor for the IngredientResult class that uses
a file handle that has the ingredient in a JSON format. This is
mostly for testing usage as I don't see a ton of practical purpose
for it in the actual application but it's helped out immensely with
for now.